+++ app/Main.hs
@@ -18,6 +18,7 @@ import Data.Set qualified as Set
 import Data.Text.IO qualified as TIO
 import Data.Version (showVersion)
 import Distribution.ModuleName (ModuleName)
+import Distribution.Types.PackageName (PackageName)
 import Language.Haskell.TH.Env (envQ)
 import Options.Applicative
 import Ormolu
@@ -392,7 +393,8 @@ parseFixityDeclaration :: ReadM [(OpName, FixityInfo)]
 parseFixityDeclaration = eitherReader parseFixityDeclarationStr
 
 -- | Parse a module reexport declaration.
-parseModuleReexportDeclaration :: ReadM (ModuleName, NonEmpty ModuleName)
+parseModuleReexportDeclaration ::
+  ReadM (ModuleName, NonEmpty (Maybe PackageName, ModuleName))
 parseModuleReexportDeclaration = eitherReader parseModuleReexportDeclarationStr
 
 -- | Parse 'ColorMode'.